requirements.txt
python-dotenv
Flask
pandas
mysql-connector
plotly-express
cryptography
plotly